backend/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
requests
python-dotenv
pydantic
